Add Rock, Paper, Scissors
Add a function to generate the computes's choice.
Add a function for the user to enter his choice (case insensitive).
Add an error-message if the user didn't choose any of the three valid choices.
Add a function to play a round and determine the winner.
Add a function to play multiple roundes and determine the winner.